refactor: Move RegisterUIState and PhoneNumberUIState to a shared package
