+++ identity_presentation/src/androidUnitTest/kotlin/net/thechance/mena/identity/presentation/screen/register/RegisterOtpViewModelTest.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.test.StandardTestDispatcher
 import net.thechance.mena.identity.domain.repository.RegisterRepository
 import net.thechance.mena.identity.helper.BaseCoroutineTest
 import net.thechance.mena.identity.presentation.screen.register.otp.RegisterOtpViewModel
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
 import org.junit.Before
 import org.junit.Test
