+++ identity_presentation/src/androidUnitTest/kotlin/net/thechance/mena/identity/presentation/screen/register/accountCreated/AccountCreatedViewModelTest.kt
@@ -11,7 +11,7 @@ import net.thechance.mena.identity.domain.repository.AuthenticationRepository
 import net.thechance.mena.identity.domain.repository.RegistrationDraftRepository
 import net.thechance.mena.identity.helper.BaseCoroutineTest
 import net.thechance.mena.identity.presentation.screen.register.shared.AuthUiState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.toPhoneNumberUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.toPhoneNumberUIState
 import org.junit.Before
 import org.junit.Test
