+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/api/IdentityFeatureApiImpl.kt
@@ -25,7 +25,7 @@ import net.thechance.mena.identity.presentation.screen.register.accountCreated.A
 import net.thechance.mena.identity.presentation.screen.register.shared.AuthUiState
 import net.thechance.mena.identity.presentation.screen.register.shared.toAuthUIState
 import net.thechance.mena.identity.presentation.screen.register.shared.toAuthUIStateJsonString
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.toPhoneNumberUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.toPhoneNumberUIState
 import net.thechance.mena.identity.presentation.screen.register.uploadProfileImage.UploadProfileImageScreen
 import org.koin.compose.koinInject
