+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/createPassword/CreatePasswordViewModel.kt
@@ -11,8 +11,8 @@ import net.thechance.mena.identity.domain.model.RegistrationDraft
 import net.thechance.mena.identity.domain.repository.RegistrationDraftRepository
 import net.thechance.mena.identity.domain.useCase.validation.mobileNumber.PasswordValidator
 import net.thechance.mena.identity.presentation.base.BaseScreenModel
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.toPhoneNumber
+import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.toPhoneNumber
 
 class CreatePasswordViewModel(
     private val passwordValidator: PasswordValidator,
