+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/datePicker/DatePickerScreenUIEffect.kt
@@ -1,6 +1,6 @@
 package net.thechance.mena.identity.presentation.screen.register.datePicker
 
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
 
 sealed interface DatePickerScreenUIEffect {
     data class NavigateToSelectGender(val registerUIState: RegisterUIState) : DatePickerScreenUIEffect
