+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/otp/RegisterOtpScreen.kt
@@ -30,10 +30,10 @@ import net.thechance.mena.identity.presentation.components.OtpInput
 import net.thechance.mena.identity.presentation.components.PageDescription
 import net.thechance.mena.identity.presentation.components.snackBar.IdentitySnackBarController
 import net.thechance.mena.identity.presentation.screen.register.enterName.EnterNameScreen
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.convertJsonStringToRegisterUIState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.toPhoneNumberUIState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.toRegisterJsonString
+import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.convertJsonStringToRegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.toPhoneNumberUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.toRegisterJsonString
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import org.koin.core.parameter.parametersOf
