+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/phoneEntry/RegisterPhoneEntryViewModel.kt
@@ -14,8 +14,8 @@ import net.thechance.mena.identity.presentation.base.errorState.ErrorState
 import net.thechance.mena.identity.presentation.mapper.mapAuthenticationErrorToMessage
 import net.thechance.mena.identity.presentation.mapper.mapErrorToMessage
 import net.thechance.mena.identity.presentation.screen.countryPicker.menaCountries.MenaCountry
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.PhoneNumberUIState
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.PhoneNumberUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
 import org.jetbrains.compose.resources.StringResource
 
 class RegisterPhoneEntryViewModel(
