+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/selectGender/SelectGenderScreen.kt
@@ -23,8 +23,8 @@ import net.thechance.mena.identity.presentation.components.PageDescription
 import net.thechance.mena.identity.presentation.components.snackBar.IdentitySnackBarController
 import net.thechance.mena.identity.presentation.screen.register.phoneEntry.RegisterPhoneEntryScreen
 import net.thechance.mena.identity.presentation.screen.register.selectGender.components.SessionExpiredDialog
+import net.thechance.mena.identity.presentation.screen.register.shared.convertJsonStringToRegisterUIState
 import net.thechance.mena.identity.presentation.screen.register.shared.toAuthUIStateJsonString
-import net.thechance.mena.identity.presentation.screen.register.shared.uiState.convertJsonStringToRegisterUIState
 import net.thechance.mena.identity.presentation.screen.register.uploadProfileImage.UploadProfileImageScreen
 import org.jetbrains.compose.resources.stringResource
 import org.koin.core.parameter.parametersOf
