+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/shared/RegisterUIState.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.register.shared.uiState
+package net.thechance.mena.identity.presentation.screen.register.shared
 
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
